+++ src/fobi/__init__.py
@@ -1,5 +1,5 @@
 __title__ = 'django-fobi'
-__version__ = '0.19'
+__version__ = '0.19.1'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2014-2022 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'