+++ docs/changelog.rst
@@ -15,6 +15,12 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.19.1
+------
+2022-07-12
+
+- Added class-based view for Dashboard.
+
 0.19
 ----
 2022-07-11